Move NetBoxFakeRequest and copy_safe_request() from utilities.utils to utilities.request
